backend/requirements.txt
fastapi
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
pydantic
requests
uvicorn
python-dotenv